chore: Refactor imports to use core module structure
