+++ src/quickapp/core/agent/_attachment_filter.py
