+++ src/quickapp/core/agent/_prompt_providers.py
