+++ src/quickapp/core/agent/agent_settings.py
