+++ src/quickapp/core/agent/message_logger.py
