+++ src/quickapp/core/application/_initialization_error_handler.py
