+++ src/quickapp/core/application/_messages_setup.py
@@ -6,7 +6,7 @@ from aidial_sdk.chat_completion import Message, Role, ToolCall
 from aidial_sdk.utils.pydantic import ExtraAllowModel
 from injector import ProviderOf, inject
 
-from quickapp.agent.models import TOOL_EXECUTION_HISTORY
+from quickapp.core.agent.models import TOOL_EXECUTION_HISTORY
 from quickapp.common.abstract.base_transformer import MessagesTransformer
 
 logger = logging.getLogger(__name__)
