+++ src/quickapp/core/application/_quick_app_application.py
@@ -8,7 +8,7 @@ from fastapi import FastAPI
 from fastapi_injector import InjectorMiddleware, RequestScopeOptions, attach_injector
 from injector import Injector, inject
 
-from quickapp.application._otel_settings import _OtelSettings
+from quickapp.core.application._otel_settings import _OtelSettings
 from quickapp.common.base_initializer import InitializerType, invoke_initializers
 from quickapp.common.dial_settings import DialSettings
