+++ src/quickapp/starters/starters_module.py
@@ -2,7 +2,7 @@ import logging
 
 from injector import Module, inject, multiprovider
 
-from quickapp.application import Configuration
+from quickapp.core.application import Configuration
 from quickapp.config.application import ApplicationConfig
 from quickapp.config.starters import ConversationStartersConfig
