+++ src/tests/unit_tests/agent_tests/test_orchestrator_deployment_initializer.py
@@ -3,8 +3,8 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from quickapp.agent._orchestrator_deployment_initializer import _OrchestratorDeploymentInitializer
-from quickapp.agent.orchestrator_capabilities import OrchestratorCapabilities
+from quickapp.core.agent._orchestrator_deployment_initializer import _OrchestratorDeploymentInitializer
+from quickapp.core.agent import OrchestratorCapabilities
 from quickapp.common.exceptions import OrchestratorInitializationException
