+++ src/tests/unit_tests/application_tests/test_exception_message_resolver.py
@@ -2,7 +2,7 @@ import httpx
 import openai
 import pytest
 
-from quickapp.application._exception_message_resolver import resolve_exception_message
+from quickapp.core.application._exception_message_resolver import resolve_exception_message
 from quickapp.common.exceptions import OrchestratorExceedMaxIterationsException
 from quickapp.dial_core_services.exceptions import (
     ToolsetForbiddenException,
