+++ src/tests/unit_tests/application_tests/test_request_context_setup.py
@@ -6,8 +6,8 @@ import pytest
 from aidial_sdk.chat_completion import Request
 from aidial_sdk.deployment.configuration import ConfigurationRequest
 
-from quickapp.application._request_context import _RequestContext
-from quickapp.application._request_context_setup import _RequestContextSetup
+from quickapp.core.application import _RequestContext
+from quickapp.core.application._request_context_setup import _RequestContextSetup
 
 _PATCH_MODEL_VALIDATE = patch(
     "quickapp.application._request_context_setup.ApplicationConfig.model_validate",
