+++ src/tests/unit_tests/config_tests/test_default_orchestrator_deployment.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import pytest
 from pydantic_core import PydanticUndefined
 
-from quickapp.agent.agent_settings import AgentSettings
+from quickapp.core.agent import AgentSettings
 from quickapp.config.application import _orchestrator_deployment_field
 from quickapp.config.dial_deployment import DialDeploymentConfig
