+++ src/tests/unit_tests/orchestrator_attachment_strategies/lazy_on_demand/test_attachment_get_content_injector.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 import pytest
 from aidial_sdk.chat_completion import Attachment, CustomContent, Message, Role
 
-from quickapp.agent.orchestrator_capabilities import OrchestratorCapabilities
+from quickapp.core.agent import OrchestratorCapabilities
 from quickapp.orchestrator_attachment_strategies.lazy_on_demand._attachment_get_content_injector import (
     _AttachmentGetContentInjector,
 )
